// src/pages/WebEditor/types.ts
export enum ElementType {
  "pure-text" = "pure-text",
  div = "div",
  button = "button",
}

export enum StyleEnum {
  width = "width",
  height = "height",
  color = "color",
  backgroundColor = "backgroundColor",
  position = "position",
  top = "top",
  right = "right",
  bottom = "bottom",
  left = "left",
}